requirements.txt
fastapi
uvicorn
qdrant-client
sentence-transformers
langchain
openai-whisper
requests